Fix condition not working with adhoc chart dependencies
This fixes the regression introduced by #20
